Add dedicated ambient heroes for research and articles
